Add the ability to reuse captured regex groups.
This feature enables you to move data between headers and/or the
request/response line and the headers. This adds four new regex
operations:

  * reqkeep
  * reqikeep
  * rspkeep
  * rspikeep

These operators will match a regex like the normal req/rsp operators,
but matching groups will be held in a special buffer. This buffer will
persist until the next *keep rule.

To use the matches in any *rep rule, reference them with the backref
marker "\kN" where N is the number of the matching group from the *keep
rule.